% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.rutaModel}
\alias{plot.rutaModel}
\title{Plot the deep features of a learned model.}
\usage{
\method{plot}{rutaModel}(model, task, ...)
}
\arguments{
\item{model}{A trained \code{rutaModel} object.}

\item{task}{A \code{rutaTask} object.}

\item{...}{Additional parameters for the plotting library}
}
\description{
Plot the deep features of a learned model.
}

